requirements.txt
SQLAlchemy==2.0.17
python-telegram-bot[all]==20.3
asyncpg==0.27.0